+++ app/src/main/java/com/rent/data/model/flight/Flight.kt
@@ -6,7 +6,7 @@ import java.time.LocalDateTime
 
 @Parcelize
 data class Flight(
-    val idRental: Int,
+    val idRental: Long,
     val time: LocalDateTime,
     val departure: Airport,
     val destination: Airport,
